requirements.txt
python-dotenv==1.0.1
python_fasthtml==0.5.3
supabase==2.8.1
